% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annot.R, R/Annot-methods.R
\name{Annot}
\alias{Annot}
\alias{Annot-class}
\alias{refCompound,Annot-method}
\alias{qrySpectra,Annot-method}
\alias{refSpectra,Annot-method}
\alias{hits,Annot-method}
\alias{infoAnnotation,Annot-method}
\alias{show,Annot-method}
\title{S4 class containing annotation results}
\usage{
\S4method{refCompound}{Annot}(x)

\S4method{qrySpectra}{Annot}(x)

\S4method{refSpectra}{Annot}(x)

\S4method{hits}{Annot}(x)

\S4method{infoAnnotation}{Annot}(x)

\S4method{show}{Annot}(object)
}
\arguments{
\item{x}{character: signature supported}

\item{object}{character: signature supported}
}
\description{
`Annot` object structures the annotation results returned by the
\code{\link{annotate}} function. It contains reference spectra (with
metadata) along with their reference compound metadata, the query spectra
(with metadata), and a hits table linking all three items; the hits table
also contains the distance metrics and other variables that may define a hit
(e.g. adduct assumed on the query spectrum).

An \code{Annot} object is created internally by the \code{\link{annotate}}
function, but its content can be accessed externally by using: \itemize{\item
Getters of the object. See next section \item \code{\link{MS2IDgui}}
function: A GUI interface that browses visually the Annot content.\item
\code{\link{export2xlsx}} function: Export all the data to an xlsx file.}

More info in the corresponding
\href{https://jmbadia.github.io/MS2ID/articles/annotate.html#annot-object}{vignette}.
}
\section{Getters}{
 To obtain the content of any Annot's slot, please use the
  following methods \itemize{ \item hits: returns a cross-reference dataframe
  containing the hits along with their proposed adducts and common masses.
  \item qrySpectra(object): returns a \code{\link[Spectra]{Spectra}} object
  \insertCite{Spectra}{MS2ID} containing the query spectra with hits. \item
  refSpectra(object): returns a \code{\link[Spectra]{Spectra}} object with
  successful reference spectra. \item refCompound(object): returns a
  dataframe containing (reference) compound metadata of successful reference
  spectra. \item infoAnnotation(object): returns the variables used on the
  annotation process }
}

\examples{

## LOAD MS2ID LIBRARY ---
## Decompress the MS2ID library that comes with MS2ID
MS2IDzipFile <- system.file("extdata/MS2IDLibrary.zip",
                            package = "MS2ID",
                            mustWork = TRUE)
library(utils)
MS2IDFolder <- dirname(unzip(MS2IDzipFile,
                             exdir = tempdir()))[1]

## SELECT QUERY SPECTRA ---
## Decompress the query mzML files that come with MS2ID
queryFile <- system.file("extdata/QRYspectra.zip",
                         package = "MS2ID",
                         mustWork = TRUE)
queryFolder <- file.path(tempdir(), "QRYspectra")
library(utils)
unzip(queryFile, exdir = queryFolder)
## ANNOTATION ---
library(MS2ID)
MS2IDlib <- MS2ID(MS2IDFolder)
annotResult <- annotate(QRYdata = queryFolder, MS2ID = MS2IDlib)

## ANNOTATION HANDLING---
#merge hits and compound info
result <- merge(x = hits(annotResult), y = refCompound(annotResult),
               by.x = "idREFcomp", by.y = "id", all.y = FALSE)
#Subset spectra and metadata considering first hit query spectra
library(Spectra)
idQRYspect_1 <- result$idQRYspect[1]
result_1 <- dplyr::filter(result, idQRYspect  == idQRYspect_1)
qrySpct_1 <- qrySpectra(annotResult)
qrySpct_1 <- qrySpct_1[qrySpct_1$id \%in\% result_1$idQRYspect]
refSpct_1 <- refSpectra(annotResult)
refSpct_1 <- refSpct_1[refSpct_1$id \%in\% result_1$idREFspect]
}
\references{
\insertRef{Spectra}{MS2ID}
}
\seealso{
\code{\link{annotate}} function, and post annotation tools
  \code{\link{MS2IDgui}} and \code{\link{export2xlsx}}.
}
\author{
Josep M. Badia \email{josepmaria.badia@urv.cat}
}
