% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{Annotate spectra against an in-house database}
\usage{
annotate(
  QRYdir,
  MS2ID,
  noiseThresh = 0.01,
  cosSimThresh = 0.8,
  massErrThresh = 20,
  cmnPrecMass = FALSE,
  cmnNeutralMass = TRUE,
  cmnPeaks = 2,
  cmnTopPeaks = 5,
  cmnPolarity = TRUE,
  db = "all",
  nature = "all",
  ...
)
}
\arguments{
\item{QRYdir}{character(1). Path of the folder containing the mzML files
(query spectra) to annotate.}

\item{MS2ID}{MS2ID object with the in-house database to use in the
annotation.}

\item{noiseThresh}{A numeric defining the threshold used in the noise
filtering of the query spectra, considered as \% intensity relative to
base peak. e.g. noiseThresh=0.01 eliminates peaks with an intensity of
less than 1\% of the base peak.}

\item{cosSimThresh}{Numeric defining the minimum cosine similarity a
pair of spectra (query and reference) must have to be returned by the
function. (TODO: see link with the definition of cossim used)}

\item{massErrThresh}{TODO}

\item{cmnPrecMass}{-Reference spectra filter- Boolean, a TRUE value limits
the reference spectra to those that have the same precursor mass as the
query spectrum.}

\item{cmnNeutralMass}{-Reference spectra filter- Boolean, a TRUE value
limits the reference spectra to those that have a neutral mass that matches
some of the plausible query neutral masses (considering the precursor
query mass and all the possible adducts, TODO: see link).}

\item{cmnPeaks}{-Reference spectra filter- Integer limiting reference
spectra to whose with at least that number of peaks in common with the
query spectrum.}

\item{cmnTopPeaks}{-Reference spectra filter- Integer limiting the
annotation to reference spectra with at least one common peak -with the
query spectrum- among their top n most intense peaks.}

\item{cmnPolarity}{-Reference spectra filter- Boolean, a TRUE value limits
the reference spectra to those with the same polarity as the query spectrum.}

\item{db}{-Reference spectra filter- Character filtering the reference
spectra by its original database (e.g. HMDB).}

\item{nature}{-Reference spectra filter- Character filtering the reference
spectra by the spectra nature. Values are restricted to 'experimental',
'insilico' or 'all'.}

\item{...}{Arguments to be passed to internal functions}
}
\value{

}
\description{
\code{annotate} returns, for every query spectrum, a list of compound
 candidates from a database by using different metrics to compare their
 spectra.
}
\examples{

}
