% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotdSpectra.R
\docType{class}
\name{AnnotdSpectra-class}
\alias{AnnotdSpectra-class}
\alias{.AnnotdSpectra}
\title{Class container of results obtained by spectra annotation}
\description{
`AnnotdSpectra` object provides a structure to the annotation results of a
query spectra  against a  reference spectra library. It must contain
reference spectra (with metadata) along with their reference compound
metadata, the query spectra (with metadata), and a hits table linking all
three items; the hits table also contains the distance metrics and other
variables that may define a hit (e.g. adduct assumed on the query spectrum)
}
\details{
`AnnotdSpectra` objects should be created using the constructor function
`AnnotdSpectra` providing required spectra as a spectra object
(TODO:Reference).
}
\section{Slots}{

\describe{
\item{\code{hits}}{cross-reference dataframe containing the hits along with their
proposed adducts and common masses}

\item{\code{refCompound}}{dataframe containing (reference) compound metadata}

\item{\code{qrySpectra}}{spectra(n) with the query spectra}

\item{\code{refSpectra}}{spectra(n) with reference spectra}

\item{\code{.properties}}{inner info}

\item{\code{infoAnnotation}}{variables used on the annotation process}
}}

