% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMS2ID.R
\name{createMS2ID}
\alias{createMS2ID}
\title{Create a MS2ID object}
\usage{
createMS2ID(
  name = "MS2ID",
  path = ".",
  cmpdb,
  noiseThresh = 0.01,
  calcSplash = TRUE,
  calcMmi = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{name}{character(1) name of the directory where the files will be saved.}

\item{path}{character(1) with the directory's path.}

\item{cmpdb}{\code{\linkS4class{CompDb}}(1) object (with MS/MS spectra).}

\item{noiseThresh}{A numeric defining the threshold used in the noise
filtering of the MS/MS spectra. e.g. noiseThresh = 0.01 removes peaks with
an intensity less than 1\% of the base peak.}

\item{calcSplash}{boolean(1) indicating if SPLASH values (when missing) must be calculated using the \code{\link[splashR]{getSplash}} function.}

\item{calcMmi}{boolean(1) indicating if compound monoisotopic mass values (when missing) must be calculated using the \code{\link[enviPat]{check_chemform}} function.}

\item{overwrite}{boolean(1) indicating if the function can overwrite     results.}
}
\value{
a character(1) naming the resulting directory (+ path)
}
\description{
\code{createMS2ID} processes the data contained in a
\code{\linkS4class{CompDb}} object and saves the result as SQL and bigMemory
files in a directory. Later on, that directory can be load
as \linkS4class{MS2ID} object in order to be used repeatedly as reference
library with the \code{\link{annotate}} function.
}
\seealso{
\code{\linkS4class{MS2ID}}
}
