% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra-utility-functions.R
\name{.binSpectra}
\alias{.binSpectra}
\title{Bin spectra}
\usage{
.binSpectra(spectra, decimals2bin)
}
\arguments{
\item{spectra}{list where every item is a spectrum. Every spectrum is a
matrix with two rows (named 'mass-charge' and 'intensity') and a column
for every mass.}

\item{decimals2bin}{integer or numeric indicating the number of decimal places to be used on the round.}
}
\value{
The spectra list binned
}
\description{
'.binSpectra' rounds mz masses of spectra and merge those with the same
resulting value.
}
