% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customValues.R
\docType{data}
\name{.colCons}
\alias{.colCons}
\title{Vector of CONS colors}
\format{
An object of class \code{character} of length 15.
}
\usage{
.colCons
}
\description{
palette for spectra that forms consensus (MS2IDgui, CONS plot)
}
\keyword{datasets}
