% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparableSpectra.R
\name{.comparableSpectra}
\alias{.comparableSpectra}
\title{Obtain reference spectra comparable to a query spectrum}
\usage{
.comparableSpectra(QRYmz, cmnPeaks = 2, cmnTopPeaks = 5)
}
\arguments{
\item{QRYmz}{numeric vector with mz values, presumably from a query spectrum.}

\item{cmnPeaks}{Filter by fragment argument. Resulting REF spectra must
have at least a number N (cmnPeaks value) of common fragments
 with UNK spectra.}

\item{cmnTopPeaks}{REF library subsetting argument. Only REF spectra with
spectra whose top N (topMassesNum value) fragments share at list one mz 
with UNK fragments.}
}
\value{
A numeric vector with the id number of the reference spectra that 
meet the requirements.
}
\description{
This function identifies reference spectra that have peaks in common with a 
query spectrum by consulting a mz index. Arguments allow modulate 
requirements of minimum number of common peaks (cmnPeaks) or their 
relative weigth (cmnTopPeaks).
}
