% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customValues.R
\docType{data}
\name{.MS2IDcols}
\alias{.MS2IDcols}
\title{Vector of custom colors}
\format{
An object of class \code{character} of length 4.
}
\usage{
.MS2IDcols
}
\description{
Vector of custom colors
}
\keyword{datasets}
