% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra-utility-functions.R
\name{.pruneSpectra}
\alias{.pruneSpectra}
\title{Remove invalid spectra}
\usage{
.pruneSpectra(DB)
}
\arguments{
\item{DB}{object with the structure of .loadSpectra() return}
}
\value{
The db object with any invalid spectra removed, from the metadata
but also from the list of matrices
}
\description{
'.pruneSpectra' checks all spectra matrices and remove the invalid ones from
the list of matrices and metadata. A spectra matrix is considered valid if it
is a matrix with two rows and at list one column.
}
