% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2ID.R
\docType{class}
\name{MS2ID-class}
\alias{MS2ID-class}
\alias{.MS2ID}
\title{S4 class containing a spectra library}
\description{
`MS2ID` objects provide a structure to encapsulate ready-to-use MS2 spectra
database along with its metadata. This internal structure allows handling
big spectra databases and annotating query spectra at high speed and low
RAM requirement.
The metadata are stored in a SQLite database; the spectra (i.e.
peaks matrices) and a mz index in bigmemory external files.
}
\section{Slots}{

\describe{
\item{\code{dbcon}}{DBIConnection to metadata and pointers to big.matrix
data}

\item{\code{spectracon}}{a big.matrix object containing spectra}

\item{\code{mzIndexcon}}{a big.matrix object containing a mass-charge index of the
spectra}

\item{\code{.properties}}{inner info}
}}

