% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annot.R
\name{Annot}
\alias{Annot}
\alias{Annot-class}
\title{S4 class containing annotation results}
\description{
`Annot` object structures the annotation results. It must contain
reference spectra (with metadata) along with their reference compound
metadata, the query spectra (with metadata), and a hits table linking all
three items; the hits table also contains the distance metrics and other
variables that may define a hit (e.g. adduct assumed on the query spectrum)
}
\section{Getters}{
 To obtain the content of any Annot's slot, please
  use the following methods \itemize{ \item hits(object): returns a
  cross-reference dataframe containing the hits along with their proposed
  adducts and common masses. \item qrySpectra(object): returns a
  \code{\link[Spectra]{Spectra}} object (see
  \href{https://www.bioconductor.org/packages/release/bioc/html/Spectra.html}{Spectra
   package}) containing the query spectra with hits. \item
  refSpectra(object): returns a \code{\link[Spectra]{Spectra}} object with
  successful reference spectra. \item refCompound(object): returns a
  dataframe containing (reference) compound metadata of successful reference
  spectra. \item infoAnnotation(object): returns the variables used on the
  annotation process }
}

