% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export2xlsx.R
\name{export2xlsx}
\alias{export2xlsx}
\title{Export an annotation object as xlsx file}
\usage{
export2xlsx(anRslt, summarizeHits, ...)
}
\arguments{
\item{anRslt}{Annot object with the results to be exported}

\item{noCmnNeutralMassDist}{numeric(1) with a distance threshold to subset Non common neutral mass results.}

\item{metric}{char(1) with the name of the metric that must subset (according to the distance threshold) and order the identifications.}

\item{file}{char(1) name (with path) of the exported file}

\item{overwrite}{boolean(1) If TRUE, overwrite any existing file.}
}
\description{
Export an annotation object as xlsx file
}
