% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2ID.R
\name{MS2ID}
\alias{MS2ID}
\alias{MS2ID-class}
\title{In-house database for high-throughput MS/MS annotation}
\usage{
MS2ID(dir)
}
\description{
The `MS2ID` class encapsulates a MS2 spectra database along with its metadata.
Its internal structure allows to annotate query spectra - with the
\code{\link{annotate}} function - using big spectra databases at high speed
and low RAM requirements (typically 100 query spectra/min against a 1M5
spectra library). MS2ID class uses a SQLite database with the metadata and
bigmemory files to  store the spectra (i.e. peaks matrices) and their
mass-charge index.
}
\details{
`MS2ID` objects should be created using the constructor function `MS2ID`
providing the path where the library files are (i.e. SQL file
name or any bigmemory file names)
}
\section{General functions}{
 \itemize{ \item `MS2ID()`: Constructor of a MS2ID
 object. \item `annotate()`: function to annotate query spectra against a
 MS2ID library }
}

\author{
Josep M. Badia
}
