% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2ID.R
\docType{package}
\name{MS2ID}
\alias{MS2ID}
\title{High-throughput MS/MS annotation with an unique in house database}
\usage{
MS2ID(x)
}
\arguments{
\item{x}{For `MS2ID`: `character(1)` defining the directory path containing
the MS2ID db files: 'metadataDB.db',}
}
\description{
`MS2ID` objects provide a structure to encapsulate ready-to-use MS2 spectra
database along with its metadata.  This internal structure allows handling
big spectra databases and annotating query spectra at high speed and low
RAM requirement.
The metadata are stored in a SQLite database, the spectra (i.e.
peaks matrices) in bigmemory external files.
}
\details{
`MS2ID` objects should be created using the constructor function
`MS2ID` providing any file name (with path) related to the database
(i.e. SQL file name or any bigmemory file names)
}
\section{General functions}{


- `MS2ID`: connect to a compound database.
}

\seealso{
[annotate()] for the function to annotate query spectra against the
 MS2ID database
}
\author{
Josep M. Badia
}
